util: Move `metrics.go` into its own package.
In our packaging style we do follow layering of packages.
Currently, util package depends on other util/* packages.
Hence, either the dependencies should move into util
independently OR we should move metrics.go into another
package util/metrics. The latter is done since the former
changes multiple other packages and those changes do not make
sense.

Fixes #95
