+++ pkg/util/metrics/doc.go
@@ -1,0 +1,4 @@
+/*
+Package metrics contains utilities for handling status metrics.
+*/
+package metrics
