+++ pkg/util/metrics/serialize.go
@@ -1,4 +1,4 @@
-package util
+package metrics
 
 import (
 	"time"
